Further refinement of domain classes and storage of proper event values (AISAB-14).

+++ ais-ab-event-db/src/main/java/dk/dma/ais/abnormal/event/db/domain/Position.java
@@ -35,10 +35,10 @@ public class Position implements Comparable<Position> {
     private long id;
 
     @NotNull
-    private float latitude;
+    private double latitude;
 
     @NotNull
-    private float longitude;
+    private double longitude;
 
     @NotNull
     @Past
@@ -47,19 +47,19 @@ public class Position implements Comparable<Position> {
     public Position() {
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
